all builtins fixed for taking t_command struct as param
