+++ Makefile
@@ -5,10 +5,11 @@ SOURCES = \
 		src/handling_quotes.c \
 		src/handling_redir.c \
 		src/parsing.c \
-		src/tokens_to_array.c \
-		src/main.c
-		#src/builtins/exec_builtin.c \
-		#src/execution/test_main.c 
+		src/builtins/exec_builtin.c \
+		src/builtins/enviroment.c \
+		src/builtins/env_utils.c \
+		src/execution/test_main.c 
+		#src/main.c 
 
 OBJECTS = $(SOURCES:.c=.o)
 
@@ -16,7 +17,7 @@ LIBFT = ./libft/libft.a
 
 CC = cc
 
-CFLAGS += -Wall -Wextra -Werror
+CFLAGS += -Wall -Wextra -Werror -g
 
 RM = rm -f
