+++ src/execution/test_main.c
@@ -6,7 +6,7 @@
 /*   By: djelacik <djelacik@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/02 18:23:55 by djelacik          #+#    #+#             */
-/*   Updated: 2024/11/12 21:02:56 by djelacik         ###   ########.fr       */
+/*   Updated: 2024/11/13 12:17:59 by djelacik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ int	main(int argc, char **argv, char **envp)
 	command = malloc(sizeof(t_command));
 	command->token_count = 3;
 	command->tokens = malloc(sizeof(t_tokens) * command->token_count);
-	command->tokens[0].token_string = ft_strdup("TEST=");
-	command->tokens[1].token_string = ft_strdup("TEST=2");
-	command->tokens[2].token_string = ft_strdup("TEST=3");
+	command->tokens[0].token_string = ft_strdup("TEST1");
+	command->tokens[1].token_string = ft_strdup("TEST2=");
+	command->tokens[2].token_string = ft_strdup("TEST3=3");
 	
 	if (argc < 2)
 		return (EXIT_FAILURE);
@@ -31,16 +31,19 @@ int	main(int argc, char **argv, char **envp)
 	if (ft_strcmp(argv[1], "pwd") == 0)
 		exec_pwd();
 	init_list(&head, envp);
-	//ft_env(head);
 	if (ft_strcmp(argv[1], "unset") == 0)
 	{
 		ft_unset(&head, argv[2]);
 		ft_env(head);
 	}
+	export_add(&head, command);
 	if (ft_strcmp(argv[1], "export") == 0)
 	{
-		export_add(&head, command);
+		export_print(head);
 	}
-	ft_env(head);
+	if (ft_strcmp(argv[1], "env") == 0)
+	{
+		ft_env(head);
+	}
 	return (EXIT_SUCCESS);
 }
